v1.3.39 follow-up: Gemma 4 multi-turn auto-bypass + RotatingKVCache meta fix
- scheduler + mllm_scheduler: detect mixed-attention models (Gemma 4 =
  sliding + full) at init and transparently force _bypass_prefix_cache=True
  on every request. Cures "step-by-step-step-by-step…" word-loops at
  rep_pen=1.1 without touching any other model family. Confirmed live on
  Gemma-4-26B-JANG_4M and Gemma-4-31B-JANG_4M.

- scheduler._rebuild_meta_state_after_truncation: helper that preserves
  (keep, max_size) when truncating RotatingKVCache KV for the paged store.
  Previously slot 0 was overwritten under the assumption it was 'offset'
  (true for KVCache, false for RotatingKVCache). Refuses to store when a
  circular buffer has already wrapped.

- mllm_scheduler._truncate_hybrid_cache: preserve RotatingKVCache class
  instead of demoting to plain KVCache — sliding-window layers keep their
  window state on reconstruct.

- prefix_cache + scheduler: bf16 numpy round-trip now casts through fp32
  (not fp16) so attention KV precision survives the Metal-safety detour.

- tests/test_cache_bypass.py: +9 source-level tests for mixed-attention
  detection + RotatingKVCache meta_state preservation + wrapped-buffer
  refusal. 37/37 pass.

Live-validated on 8 models, 56 multi-turn cache turns, zero regressions:
  - Qwen3-0.6B-8bit (stock MLX)           7/7 ✅
  - Qwen3.5-4B-JANG_4K                    7/7 ✅
  - Qwen3.5-VL-4B-JANG_4S                 7/7 ✅ (VL + image alternation)
  - Nemotron-Cascade-2-30B-JANG_2L        7/7 ✅ (hybrid SSM)
  - Gemma-4-26B-JANG_4M                   7/7 ✅ (auto-bypass active)
  - Qwen3.5-122B-A10B-JANG_2S             7/7 ✅ (VL + hybrid SSM)
  - Gemma-4-31B-JANG_4M                   7/7 ✅ (auto-bypass active)
  - MiniMax-M2.5-JANG_2L                  7/7 ✅ (thinking model)

+++ vmlx_engine/prefix_cache.py
@@ -1246,17 +1246,26 @@ class BlockAwarePrefixCache:
                                 v_w, v_s, v_b, group_size=g_size, bits=q_bits,
                             )
                             if 'bfloat16' in str(k_dq.dtype):
-                                k_dq = k_dq.astype(mx.float16)
-                                v_dq = v_dq.astype(mx.float16)
+                                # Use float32 (not float16) for the numpy round-trip.
+                                # fp16 has only 5 exponent bits vs bf16's 8 — casting
+                                # down silently clips/loses precision for many
+                                # attention KV values. On Gemma 4 JANG this caused
+                                # "step-by-step" word loops on the 3rd multi-turn
+                                # request because the cached KV drift pushed the
+                                # sampler into a degenerate rep_pen-proof basin.
+                                k_dq = k_dq.astype(mx.float32)
+                                v_dq = v_dq.astype(mx.float32)
                             mx.eval(k_dq, v_dq)
                             np_sources[idx] = (np.array(k_dq), np.array(v_dq), k_dq.dtype)
                             continue
                         if hasattr(keys, 'shape'):
                             k_np, v_np = keys, values
-                            # numpy doesn't support bfloat16 — cast first
+                            # numpy doesn't support bfloat16 — cast through fp32
+                            # (NOT fp16, which silently clips). See the Gemma 4
+                            # multi-turn word-loop note above.
                             if hasattr(k_np, 'dtype') and 'bfloat16' in str(k_np.dtype):
-                                k_np = k_np.astype(mx.float16)
-                                v_np = v_np.astype(mx.float16)
+                                k_np = k_np.astype(mx.float32)
+                                v_np = v_np.astype(mx.float32)
                             np_sources[idx] = (np.array(k_np), np.array(v_np), keys.dtype)
                     except Exception as _npe:
                         logger.debug(f"np_sources skip layer {idx}: {_npe}")
